fix error on deleting non-existing package file
Former-commit-id: a7cbf94bb4709f1714b6673aea65bcaecf55ee32

+++ build/deployment-functions.ps1
@@ -7,8 +7,10 @@ function PublishZipPackage($SourceFolder, $TargetFile) {
     Write-Host "##teamcity[blockOpened name='Publishing zip package artifact']"
     Write-Host "##teamcity[progressStart 'Publishing zip package artifact']"
 
-    Remove-Item $TargetFile
-
+	IF (Test-Path $TargetFile){
+		Remove-Item $TargetFile
+	}	
+    
     [Reflection.Assembly]::LoadWithPartialName("System.IO.Compression.FileSystem")
     [System.AppDomain]::CurrentDomain.GetAssemblies()
